requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-core==1.30.1
azure-eventhub==5.11.6
azure-servicebus==7.11.4
blinker==1.7.0
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
dataclasses==0.6
deprecation==2.1.0
Flask==3.0.2
Flask-RESTful==0.3.10
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==2.10
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==23.2
pluggy==1.4.0
postgrest==0.16.1
pydantic==2.6.3
pydantic_core==2.16.3
pytest==7.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.2
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
toml==0.10.2
typing_extensions==4.10.0
urllib3==1.26.18
websockets==11.0.3
Werkzeug==3.0.1
